Move Observation tests from DataFrameSuite to DatasetSuite
